requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.18.0
numpy>=1.24.0
pandas-ta>=0.3.14b0
statsmodels>=0.14.0
yfinance>=0.2.31
futu-api>=7.1.3308
tensorflow>=2.13.0
curl-cffi>=0.5.0
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
matplotlib>=3.5.0 